refactor: reorganized elements to their respective modified feature-sliced organization
