+++ src/app/(auth)/login/page.tsx
@@ -1,5 +1,5 @@
 import { GalleryVerticalEnd } from "lucide-react";
-import { LoginForm } from "./components/login-form";
+import { LoginForm } from "@/features/auth/components/login-form";
 export default function LoginPage() {
   return (
     <div className="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
