+++ src/app/(auth)/signup/page.tsx
@@ -1,5 +1,5 @@
 import { GalleryVerticalEnd } from "lucide-react";
-import { SignUpForm } from "./components/signup-form";
+import { SignUpForm } from "@/features/auth/components/signup-form";
 export default function SignupPage() {
   return (
     <div className="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
